hello_langgraph/pyproject.toml
[project]
name = "hello-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.1.3",
    "langgraph-cli>=0.4.19",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
]